Adds lhctl execute userTaskRun and lhctl assign userTaskRun
